+++ pointers/pointers.c
@@ -1,53 +1,5 @@
 #include <stdio.h>
-
-/*
-    Swaps the integer values being pointed at by a and b. Keep in
-    mind when you need to access a pointer's actual value (the 
-    address it's referring to) or the value at the address it's 
-    pointing at.
-*/
-void swap(int* a, int* b)
-{
-
-}
-
-/*
-    Given a character pointer s, returns the number of characters
-    it is pointing to. Think about how pointer arithmetic can 
-    help you with this.
-*/
-int string_length(char *s)
-{
-
-}
-
-/*
-    Given an empty (NULL) character pointer x and a character pointer y,
-    copies the character contents of y over to x. Again, pointer arithmetic
-    is necessary here. Also, make sure x points to a null character at its 
-    end to terminate it properly. 
-*/
-void string_copy(char *x, char *y)
-{
-
-}
-
-/* 
-    Compares the character strings m and n and returns negative,
-    0, or positive if n is lexicographically less than, equal to,
-    or greater than n. To calculate lexicographic difference, find
-    the difference between the first characters in m and n that differ.
-    
-    For example, given matching strings, this function should 
-    return 0. Given strings m = "hello world" and n = "goodbye",
-    this function should return a negative value. Given strings
-    m = "aardvark" and n = "zebra", should return a positive
-    value.
-*/
-int string_compare(char *m, char *n)
-{
-
-}
+#include <pointers.h>
 
 #ifndef TESTING
 int main(void)
@@ -55,7 +7,9 @@ int main(void)
     int x = 10, y = 20;
     swap(&x, &y);
     printf("x=%d, y=%d\n", x, y);
-
+    
+    // char string[] = "bloop";
+    // printf("Testing\n strlen >> %i", string_length(string));
     char *hello = "Hello";
     char *world = "World";
     char buffer[1024];
