+++ pointers/pointers.h
@@ -1,12 +1,77 @@
+#include <stdio.h>
+
 #ifndef pointers_h
 #define pointers_h
 
-void swap(int *a, int *b);
+/*
+    Swaps the integer values being pointed at by a and b. Keep in
+    mind when you need to access a pointer's actual value (the 
+    address it's referring to) or the value at the address it's 
+    pointing at.
+*/
+void swap(int *a, int *b)
+{
+  int temp = *a;
+  *a = *b;
+  *b = temp;
+}
 
-int string_length(char *s);
+/*
+    Given a character pointer s, returns the number of characters
+    it is pointing to. Think about how pointer arithmetic can 
+    help you with this.
+*/
+int string_length(char *s)
+{
+  int i = 0;
+  for (; *s != '\0'; i++, s++)
+  {
+  }
+  return i;
+}
 
-void string_copy(char *a, char *b);
+/*
+    Given an empty (NULL) character pointer x and a character pointer y,
+    copies the character contents of y over to x. Again, pointer arithmetic
+    is necessary here. Also, make sure x points to a null character at its 
+    end to terminate it properly. 
+*/
 
-int string_compare(char *m, char *n);
+void string_copy(char *x, char *y)
+{
+  while (*y != '\0')
+  {
+    *x = *y;
+    x++;
+    y++;
+  }
+  *x = '\0';
+  printf("Called string_copy %s", x);
+}
 
+/* 
+    Compares the character strings m and n and returns negative,
+    0, or positive if n is lexicographically less than, equal to,
+    or greater than n. To calculate lexicographic difference, find
+    the difference between the first characters in m and n that differ.
+    
+    For example, given matching strings, this function should 
+    return 0. Given strings m = "hello world" and n = "goodbye",
+    this function should return a negative value. Given strings
+    m = "aardvark" and n = "zebra", should return a positive
+    value.
+*/
+
+int string_compare(char *m, char *n)
+{
+  printf("Called strings_compare");
+  for (; *m == *n; m++, n++)
+  {
+    if (*m == '\0')
+    {
+      return 0;
+    }
+  }
+  return *m - *n;
+}
 #endif
